added stack split, comments, cleaned up actions files

+++ fitg/backend/actions/movement.py
@@ -50,30 +50,47 @@ def merge_stack(session, src_id, des_id):
     src_stack = session.query(Stack).filter_by(id = src_id).one()
     des_stack = session.query(Stack).filter_by(id = des_id).one()
 
-    #assert isinstance(src_stack, Stack)
-    #assert isinstance(des_stack, Stack)
-
     # check if stacks are same team? same location?
 
-    assert src_stack.location == des_stack.location
+    try:
+        assert src_stack.location == des_stack.location
+        assert src_stack.characters[0].side == des_stack.characters[0].side
+    except:
+        success = False
+    else:
+        success = True
+        for character in src_stack.characters:
+            character.stack_id = des_id
+        for unit in src_stack.militaryunits:
+            unit.stack_id = des_id
 
-    for character in src_stack.characters:
-        character.stack_id = des_id
-    for unit in src_stack.militaryunits:
-        unit.stack_id = des_id
+    if success:
+        session.add(des_stack)
+        session.delete(src_stack)
+        session.commit()
+        return success
+    else:
+        return succes, "FATAL: Stacks cannot be merged"
 
-    session.add(des_stack)
-    session.delete(src_stack)
-    session.commit()
+def split_stack(session, src_id, unit_id, is_character):
+    src_stack = session.query(Stack).filter_by(id = src_id).one()
 
-if __name__ == "__main__":
-    loadDatabase()
+    try:
+        if is_character:
+            moving_unit = session.query(Character).filter_by(id = unit_id).one()
+        else:
+            moving_unit = session.query(Unit).filter_by(id = unit_id).one()
+    except:
+        success = False
+    else:
+        success = True
+        new_stack = Stack()
+        session.add(new_stack, moving_unit)
+        moving_unit.stack = new_stack
+        
 
-    #move_environ(1,'3110')
-    #move_environ(2,'3112')
-
-    session = Session()
-    print session.query(Stack).filter_by(id = 1).one().location
-    
-    merge_stack(1,2)
-    print session.query(Stack).filter_by(id = 2).one().characters
+    if success:
+        session.commit()
+        return success, new_stack.__dict__
+    else:
+        return success, "FATAL: Unable to split stack"
